Move test util
