+++ backend/grants_management_api/tests/auth/test_authorization_enforcer.py
@@ -1,21 +1,17 @@
 import pytest
 from apiflask import HTTPError
-from grants_shared.adapters import db
 
 from src.auth.authorization_enforcer import AuthorizationEnforcer
 from src.constants.lookup_constants import MgmtPrivilege, MgmtResourceType
-from src.db.models.resource_models import AbstractResourceTableMixin, MgmtRole
-from src.db.models.user_models import MgmtUser
 from tests.db.models.factories import (
     DepartmentFactory,
     MgmtInternalResourceFactory,
-    MgmtResourceUserFactory,
-    MgmtResourceUserRoleFactory,
     MgmtRoleFactory,
     MgmtUserFactory,
     SubagencyFactory,
     TeamFactory,
 )
+from tests.test_utils.auth_test_utils import setup_user_with_roles
 
 ######################################
 # Resource Fixtures
@@ -91,45 +87,6 @@ def internal_resource2(enable_factory_create):
 ######################################
 # Tests
 ######################################
-
-# TODO - move this somewhere
-
-
-def setup_user_with_roles(
-    db_session: db.Session,
-    resources: list[AbstractResourceTableMixin],
-    user: MgmtUser | None = None,
-    *,
-    roles: list[MgmtRole] | None = None,
-    privileges: list[MgmtPrivilege] | None = None,
-):
-    if user is None:
-        user = MgmtUserFactory.create()
-
-    if roles is None and privileges is None:
-        raise Exception("One of roles or privileges is required for setup_user_with_roles")
-    if roles is not None and privileges is not None:
-        raise Exception("Exactly one of roles or privileges is required for setup_user_with_roles")
-
-    resource_types = {r.get_resource_type() for r in resources}
-
-    # If privileges were passed in, use those to make a role
-    if privileges is not None:
-        roles = [MgmtRoleFactory.create(privileges=privileges, resource_types=resource_types)]
-
-    # Create a connection to every resource passed in
-    for resource in resources:
-
-        resource_user = MgmtResourceUserFactory.create(
-            mgmt_resource=resource.resource, mgmt_user=user, resource_user_roles=[]
-        )
-        for role in roles:
-            MgmtResourceUserRoleFactory.create(mgmt_resource_user=resource_user, mgmt_role=role)
-
-    # Make any subsequent calls with these objects go back to the DB.
-    db_session.expire_all()
-
-    return user
 
 
 def test_user_with_no_roles_cannot_access_anything(
